pyproject.toml
[project]
name = "tars"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.19",
    "langgraph>=0.4.8",
    "pydantic>=2.11.5",
    "python-telegram-bot[job-queue]>=22.1",
    "pytz>=2025.2",
    "ruff>=0.11.13",
]
